python_server/requirements.txt
setuptools
grpcio
grpcio-tools
psutil
websockets
